adding dummy examples for custom marines
saving my progress
