+++ src/Addons/34thPRC_ArmourCustom/data/marines2/player/demo-player/urban/demo-player_marine_shoulders.paa
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:ef801ebc706a85ccde6bc9fbc6326eafd5d4c0161fec3d7b59d9c7c3323e52eb
+size 1359323
