+++ src/Addons/34thPRC_ArmourCustom/data/marines2/player/demo-player/urban/demo-player_marine_straps.paa
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:124250ff62f627e59c3188e58f7aa96b4d99a68bf64f6b00657f413c38ec1aca
+size 1106466
