+++ src/Addons/34thPRC_ArmourCustom/data/odst2/yandere/config_weapons.hpp
@@ -1,2084 +1,5 @@
-//Yandere Custom, Helmets -> Armors
-class 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Urban : 34thPRC_ArmourStandard_ODST2_CH252D_SilverVisor_Urban
-{
-	displayName="[34th] CH252D Urban 'Yandere' (Custom Visor)";
-	author="Over Yandere";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_helmet_urban.paa",
-		"34thPRC_ArmourCustom\data\odst2\yandere\visor\yandere_odst_visor.paa"
-	};
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_SilverVisor_Urban : 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Urban
-{
-    displayName="[34th] CH252D Urban 'Yandere' (Silver Visor)";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_helmet_urban.paa",
-        "34thPRC_ArmourStandard\data\odst2\visor\34th_ODST_Silvervisor.paa"
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_RedVisor_Urban : 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Urban
-{
-	displayName="[34th] CH252D Urban 'Yandere' (Red Visor)";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_helmet_urban.paa",
-        "34thPRC_ArmourStandard\data\odst2\visor\34th_ODST_Redvisor.paa"
-    };
-}
-class 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_GoldVisor_Urban : 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Urban
-{
-	displayName="[34th] CH252D Urban 'Yandere' (Gold Visor)";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_helmet_urban.paa",
-        "34thPRC_ArmourStandard\data\odst2\visor\34th_ODST_Goldvisor.paa"
-    };
-}
-class 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Arid : 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Urban
-{
-	displayName="[34th] CH252D Arid 'Yandere' (Custom Visor)";
-	author="Over Yandere";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_helmet_arid.paa",
-		"34thPRC_ArmourCustom\data\odst2\yandere\visor\yandere_odst_visor.paa"
-	};
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_SilverVisor_Arid : 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Urban
-{
-    displayName="[34th] CH252D Arid 'Yandere' (Silver Visor)";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_helmet_arid.paa",
-        "34thPRC_ArmourStandard\data\odst2\visor\34th_ODST_Silvervisor.paa"
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_RedVisor_Arid : 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Urban
-{
-    displayName="[34th] CH252D Arid 'Yandere' (Red Visor)";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_helmet_arid.paa",
-        "34thPRC_ArmourStandard\data\odst2\visor\34th_ODST_Redvisor.paa"
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_GoldVisor_Arid : 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Urban
-{
-    displayName="[34th] CH252D Arid 'Yandere' (Gold Visor)";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_helmet_arid.paa",
-        "34thPRC_ArmourStandard\data\odst2\visor\34th_ODST_Goldvisor.paa"
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Arctic : 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Urban
-{
-	displayName="[34th] CH252D Arctic 'Yandere' (Custom Visor)";
-	author="Over Yandere";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_helmet_arctic.paa",
-		"34thPRC_ArmourCustom\data\odst2\yandere\visor\yandere_odst_visor.paa"
-	};
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_SilverVisor_Arctic : 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Urban
-{
-    displayName="[34th] CH252D Arctic 'Yandere' (Silver Visor)";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_helmet_arctic.paa",
-        "34thPRC_ArmourStandard\data\odst2\visor\34th_ODST_Silvervisor.paa"
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_RedVisor_Arctic : 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Urban
-{
-	displayName="[34th] CH252D Arctic 'Yandere' (Red Visor)";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_helmet_arctic.paa",
-        "34thPRC_ArmourStandard\data\odst2\visor\34th_ODST_Redvisor.paa"
-    };
-}
-class 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_GoldVisor_Arctic : 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Urban
-{
-	displayName="[34th] CH252D Arctic 'Yandere' (Gold Visor)";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_helmet_arctic.paa",
-        "34thPRC_ArmourStandard\data\odst2\visor\34th_ODST_Goldvisor.paa"
-    };
-}
-class 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Tropic : 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Urban
-{
-	displayName="[34th] CH252D Tropic 'Yandere' (Custom Visor)";
-	author="Over Yandere";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_helmet_tropic.paa",
-		"34thPRC_ArmourCustom\data\odst2\yandere\visor\yandere_odst_visor.paa"
-	};
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_SilverVisor_Tropic : 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Urban
-{
-    displayName="[34th] CH252D Tropic 'Yandere' (Silver Visor)";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_helmet_tropic.paa",
-        "34thPRC_ArmourStandard\data\odst2\visor\34th_ODST_Silvervisor.paa"
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_RedVisor_Tropic : 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Urban
-{
-	displayName="[34th] CH252D Tropic 'Yandere' (Red Visor)";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_helmet_tropic.paa",
-        "34thPRC_ArmourStandard\data\odst2\visor\34th_ODST_Redvisor.paa"
-    };
-}
-class 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_GoldVisor_Tropic : 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Urban
-{
-	displayName="[34th] CH252D Tropic 'Yandere' (Gold Visor)";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_helmet_tropic.paa",
-        "34thPRC_ArmourStandard\data\odst2\visor\34th_ODST_Goldvisor.paa"
-    };
-}
-class 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Woodland : 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Urban
-{
-	displayName="[34th] CH252D Woodland 'Yandere' (Custom Visor)";
-	author="Over Yandere";
-	hiddenSelectionsTextures[]=
-	{
-		"34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_helmet_woodland.paa",
-		"34thPRC_ArmourCustom\data\odst2\yandere\visor\yandere_odst_visor.paa"
-	};
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_SilverVisor_Woodland : 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Urban
-{
-    displayName="[34th] CH252D Woodland 'Yandere' (Silver Visor)";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_helmet_woodland.paa",
-        "34thPRC_ArmourStandard\data\odst2\visor\34th_ODST_Silvervisor.paa"
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_RedVisor_Woodland : 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Urban
-{
-	displayName="[34th] CH252D Woodland 'Yandere' (Red Visor)";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_helmet_woodland.paa",
-        "34thPRC_ArmourStandard\data\odst2\visor\34th_ODST_Redvisor.paa"
-    };
-}
-class 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_GoldVisor_Woodland : 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Urban
-{
-	displayName="[34th] CH252D Woodland 'Yandere' (Gold Visor)";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_helmet_woodland.paa",
-        "34thPRC_ArmourStandard\data\odst2\visor\34th_ODST_Goldvisor.paa"
-    };
-}
-
-//Deploarized Helmets
-class 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Urban_dp : 34thPRC_ArmourStandard_ODST2_CH252D_SilverVisor_Urban_dp
-{
-	displayName="[34th] CH252D Urban 'Yandere' (Custom Visor)";
-    author="Over Yandere";
-	scope=0;
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_helmet_urban.paa",// Helmet Texture
-        "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Depolarized_Visor_CA.paa" // Visor Texture
-    };
-	class ItemInfo: ItemInfo
-    {
-        uniformModel="MA_Armor\data\Helmets\Human_ODST\Reach_ODST_Helm_dp.p3d";
-        hiddenSelectionsTextures[]=
-        {
-            "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Helmet_CO.paa",
-            "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Depolarized_Visor_CA.paa"
-        };
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Arid_dp : 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Urban
-{
-    displayName="[34th] CH252D Arid 'Yandere' (Custom Visor)";
-    author="Over Yandere";
-	scope=0;
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_helmet_arid.paa",// Helmet Texture
-        "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Depolarized_Visor_CA.paa" // Visor Texture
-    };
-     class ItemInfo: ItemInfo
-    {
-        uniformModel="MA_Armor\data\Helmets\Human_ODST\Reach_ODST_Helm_dp.p3d";
-        hiddenSelectionsTextures[]=
-        {
-            "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Helmet_CO.paa",
-            "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Depolarized_Visor_CA.paa"
-        };
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Arctic_dp : 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Urban
-{
-    displayName="[34th] CH252D Arctic 'Yandere' (Custom Visor)";
-    author="Over Yandere";
-	scope=0;
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_helmet_arctic.paa",// Helmet Texture
-        "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Depolarized_Visor_CA.paa" // Visor Texture
-    };
-     class ItemInfo: ItemInfo
-    {
-        uniformModel="MA_Armor\data\Helmets\Human_ODST\Reach_ODST_Helm_dp.p3d";
-        hiddenSelectionsTextures[]=
-        {
-            "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Helmet_CO.paa",
-            "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Depolarized_Visor_CA.paa"
-        };
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Tropic_dp : 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Urban
-{
-    displayName="[34th] CH252D Tropic 'Yandere' (Custom Visor)";
-    author="Over Yandere";
-	scope=0;
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_helmet_tropic.paa",// Helmet Texture
-        "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Depolarized_Visor_CA.paa" // Visor Texture
-    };
-     class ItemInfo: ItemInfo
-    {
-        uniformModel="MA_Armor\data\Helmets\Human_ODST\Reach_ODST_Helm_dp.p3d";
-        hiddenSelectionsTextures[]=
-        {
-            "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Helmet_CO.paa",
-            "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Depolarized_Visor_CA.paa"
-        };
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Woodland_dp : 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Urban
-{
-    displayName="[34th] CH252D Woodland 'Yandere' (Custom Visor)";
-    author="Over Yandere";
-	scope=0;
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_helmet_woodland.paa",// Helmet Texture
-        "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Depolarized_Visor_CA.paa" // Visor Texture
-    };
-     class ItemInfo: ItemInfo
-    {
-        uniformModel="MA_Armor\data\Helmets\Human_ODST\Reach_ODST_Helm_dp.p3d";
-        hiddenSelectionsTextures[]=
-        {
-            "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Helmet_CO.paa",
-            "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Depolarized_Visor_CA.paa"
-        };
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_SilverVisor_Urban_dp : 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Urban
-{
-	displayName="[34th] CH252D Urban 'Yandere' (Silver Visor)";
-    author="Over Yandere";
-	scope=0;
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_helmet_urban.paa",// Helmet Texture
-        "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Depolarized_Visor_CA.paa" // Visor Texture
-    };
-	class ItemInfo: ItemInfo
-    {
-        uniformModel="MA_Armor\data\Helmets\Human_ODST\Reach_ODST_Helm_dp.p3d";
-        hiddenSelectionsTextures[]=
-        {
-            "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Helmet_CO.paa",
-            "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Depolarized_Visor_CA.paa"
-        };
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_SilverVisor_Arid_dp : 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Urban
-{
-    displayName="[34th] CH252D Arid 'Yandere' (Silver Visor)";
-    author="Over Yandere";
-	scope=0;
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_helmet_arid.paa",// Helmet Texture
-        "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Depolarized_Visor_CA.paa" // Visor Texture
-    };
-     class ItemInfo: ItemInfo
-    {
-        uniformModel="MA_Armor\data\Helmets\Human_ODST\Reach_ODST_Helm_dp.p3d";
-        hiddenSelectionsTextures[]=
-        {
-            "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Helmet_CO.paa",
-            "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Depolarized_Visor_CA.paa"
-        };
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_SilverVisor_Arctic_dp : 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Urban
-{
-    displayName="[34th] CH252D Arctic 'Yandere' (Silver Visor)";
-    author="Over Yandere";
-	scope=0;
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_helmet_arctic.paa",// Helmet Texture
-        "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Depolarized_Visor_CA.paa" // Visor Texture
-    };
-     class ItemInfo: ItemInfo
-    {
-        uniformModel="MA_Armor\data\Helmets\Human_ODST\Reach_ODST_Helm_dp.p3d";
-        hiddenSelectionsTextures[]=
-        {
-            "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Helmet_CO.paa",
-            "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Depolarized_Visor_CA.paa"
-        };
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_SilverVisor_Tropic_dp : 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Urban
-{
-    displayName="[34th] CH252D Tropic 'Yandere' (Silver Visor)";
-    author="Over Yandere";
-	scope=0;
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_helmet_tropic.paa",// Helmet Texture
-        "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Depolarized_Visor_CA.paa" // Visor Texture
-    };
-     class ItemInfo: ItemInfo
-    {
-        uniformModel="MA_Armor\data\Helmets\Human_ODST\Reach_ODST_Helm_dp.p3d";
-        hiddenSelectionsTextures[]=
-        {
-            "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Helmet_CO.paa",
-            "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Depolarized_Visor_CA.paa"
-        };
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_SilverVisor_Woodland_dp : 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Urban
-{
-    displayName="[34th] CH252D Woodland 'Yandere' (Silver Visor)";
-    author="Over Yandere";
-	scope=0;
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_helmet_woodland.paa",// Helmet Texture
-        "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Depolarized_Visor_CA.paa" // Visor Texture
-    };
-    class ItemInfo: ItemInfo
-    {
-        uniformModel="MA_Armor\data\Helmets\Human_ODST\Reach_ODST_Helm_dp.p3d";
-        hiddenSelectionsTextures[]=
-        {
-            "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Helmet_CO.paa",
-            "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Depolarized_Visor_CA.paa"
-        };
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_RedVisor_Urban_dp : 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Urban
-{
-	displayName="[34th] CH252D Urban 'Yandere' (Red Visor)";
-    author="Over Yandere";
-	scope=0;
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_helmet_urban.paa",// Helmet Texture
-        "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Depolarized_Visor_CA.paa" // Visor Texture
-    };
-	class ItemInfo: ItemInfo
-    {
-        uniformModel="MA_Armor\data\Helmets\Human_ODST\Reach_ODST_Helm_dp.p3d";
-        hiddenSelectionsTextures[]=
-        {
-            "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Helmet_CO.paa",
-            "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Depolarized_Visor_CA.paa"
-        };
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_RedVisor_Arid_dp : 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Urban
-{
-    displayName="[34th] CH252D Arid 'Yandere' (Red Visor)";
-    author="Over Yandere";
-	scope=0;
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_helmet_arid.paa",// Helmet Texture
-        "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Depolarized_Visor_CA.paa" // Visor Texture
-    };
-     class ItemInfo: ItemInfo
-    {
-        uniformModel="MA_Armor\data\Helmets\Human_ODST\Reach_ODST_Helm_dp.p3d";
-        hiddenSelectionsTextures[]=
-        {
-            "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Helmet_CO.paa",
-            "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Depolarized_Visor_CA.paa"
-        };
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_RedVisor_Arctic_dp : 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Urban
-{
-    displayName="[34th] CH252D Arctic 'Yandere' (Red Visor)";
-    author="Over Yandere";
-	scope=0;
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_helmet_arctic.paa",// Helmet Texture
-        "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Depolarized_Visor_CA.paa" // Visor Texture
-    };
-     class ItemInfo: ItemInfo
-    {
-        uniformModel="MA_Armor\data\Helmets\Human_ODST\Reach_ODST_Helm_dp.p3d";
-        hiddenSelectionsTextures[]=
-        {
-            "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Helmet_CO.paa",
-            "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Depolarized_Visor_CA.paa"
-        };
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_RedVisor_Tropic_dp : 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Urban
-{
-    displayName="[34th] CH252D Tropic 'Yandere' (Red Visor)";
-    author="Over Yandere";
-	scope=0;
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_helmet_tropic.paa",// Helmet Texture
-        "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Depolarized_Visor_CA.paa" // Visor Texture
-    };
-     class ItemInfo: ItemInfo
-    {
-        uniformModel="MA_Armor\data\Helmets\Human_ODST\Reach_ODST_Helm_dp.p3d";
-        hiddenSelectionsTextures[]=
-        {
-            "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Helmet_CO.paa",
-            "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Depolarized_Visor_CA.paa"
-        };
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_RedVisor_Woodland_dp : 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Urban
-{
-    displayName="[34th] CH252D Woodland 'Yandere' (Red Visor)";
-    author="Over Yandere";
-	scope=0;
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_helmet_woodland.paa",// Helmet Texture
-        "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Depolarized_Visor_CA.paa" // Visor Texture
-    };
-     class ItemInfo: ItemInfo
-    {
-        uniformModel="MA_Armor\data\Helmets\Human_ODST\Reach_ODST_Helm_dp.p3d";
-        hiddenSelectionsTextures[]=
-        {
-            "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Helmet_CO.paa",
-            "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Depolarized_Visor_CA.paa"
-        };
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_GoldVisor_Urban_dp : 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Urban
-{
-	displayName="[34th] CH252D Urban 'Yandere' (Gold Visor)";
-    author="Over Yandere";
-	scope=0;
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_helmet_urban.paa",// Helmet Texture
-        "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Depolarized_Visor_CA.paa" // Visor Texture
-    };
-	 class ItemInfo: ItemInfo
-    {
-        uniformModel="MA_Armor\data\Helmets\Human_ODST\Reach_ODST_Helm_dp.p3d";
-        hiddenSelectionsTextures[]=
-        {
-            "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Helmet_CO.paa",
-            "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Depolarized_Visor_CA.paa"
-        };
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_GoldVisor_Arid_dp : 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Urban
-{
-    displayName="[34th] CH252D Arid 'Yandere' (Gold Visor)";
-    author="Over Yandere";
-	scope=0;
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_helmet_arid.paa",// Helmet Texture
-        "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Depolarized_Visor_CA.paa" // Visor Texture
-    };
-     class ItemInfo: ItemInfo
-    {
-        uniformModel="MA_Armor\data\Helmets\Human_ODST\Reach_ODST_Helm_dp.p3d";
-        hiddenSelectionsTextures[]=
-        {
-            "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Helmet_CO.paa",
-            "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Depolarized_Visor_CA.paa"
-        };
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_GoldVisor_Arctic_dp : 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Urban
-{
-    displayName="[34th] CH252D Arctic 'Yandere' (Gold Visor)";
-    author="Over Yandere";
-	scope=0;
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_helmet_arctic.paa",// Helmet Texture
-        "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Depolarized_Visor_CA.paa" // Visor Texture
-    };
-     class ItemInfo: ItemInfo
-    {
-        uniformModel="MA_Armor\data\Helmets\Human_ODST\Reach_ODST_Helm_dp.p3d";
-        hiddenSelectionsTextures[]=
-        {
-            "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Helmet_CO.paa",
-            "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Depolarized_Visor_CA.paa"
-        };
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_GoldVisor_Tropic_dp : 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Urban
-{
-    displayName="[34th] CH252D Tropic 'Yandere' (Gold Visor)";
-    author="Over Yandere";
-	scope=0;
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_helmet_tropic.paa",// Helmet Texture
-        "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Depolarized_Visor_CA.paa" // Visor Texture
-    };
-     class ItemInfo: ItemInfo
-    {
-        uniformModel="MA_Armor\data\Helmets\Human_ODST\Reach_ODST_Helm_dp.p3d";
-        hiddenSelectionsTextures[]=
-        {
-            "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Helmet_CO.paa",
-            "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Depolarized_Visor_CA.paa"
-        };
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_GoldVisor_Woodland_dp : 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Urban
-{
-    displayName="[34th] CH252D Woodland 'Yandere' (Gold Visor)";
-    author="Over Yandere";
-	scope=0;
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_helmet_woodland.paa",// Helmet Texture
-        "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Depolarized_Visor_CA.paa" // Visor Texture
-    };
-     class ItemInfo: ItemInfo
-    {
-        uniformModel="MA_Armor\data\Helmets\Human_ODST\Reach_ODST_Helm_dp.p3d";
-        hiddenSelectionsTextures[]=
-        {
-            "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Helmet_CO.paa",
-            "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Depolarized_Visor_CA.paa"
-        };
-    };
-};
-
-class 34thPRC_ArmourCustom_ODST2_Yandere_M52_Urban : 34thPRC_ArmourStandard_ODST2_M52_Urban
-{
-    displayName="[34th] M52 Combat Armour 'Yandere' Urban";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-    	"34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_cqb_urban.paa", //CQB Left
-    	"34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_cqb_urban.paa", //CQB Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_marksman_urban.paa", //Marksman_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_marksman_urban.paa", //Marksman_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_vest_urban.paa", //ODST_Bracer_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_vest_urban.paa", //ODST_Bracer_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_vest_urban.paa", //ODST_Chest
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_shoulders_urban.paa", //ODST_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_shoulders_urban.paa", //ODST_Right
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_M52_B_Urban : 34thPRC_ArmourStandard_ODST2_M52_B_Urban
-{
-    displayName="[34th] M52-B Combat Armour 'Yandere' Urban";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-    	"34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_cqb_urban.paa", //CQB Left
-    	"34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_cqb_urban.paa", //CQB Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_marksman_urban.paa", //Marksman_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_marksman_urban.paa", //Marksman_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_vest_urban.paa", //ODST_Bracer_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_vest_urban.paa", //ODST_Bracer_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_vest_urban.paa", //ODST_Chest
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_shoulders_urban.paa", //ODST_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_shoulders_urban.paa", //ODST_Right
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_M52_C_Urban : 34thPRC_ArmourStandard_ODST2_M52_C_Urban
-{
-	displayName="[34th] M52-C Combat Armour 'Yandere' Urban";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-    	"34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_cqb_urban.paa", //CQB Left
-    	"34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_cqb_urban.paa", //CQB Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_marksman_urban.paa", //Marksman_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_marksman_urban.paa", //Marksman_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_vest_urban.paa", //ODST_Bracer_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_vest_urban.paa", //ODST_Bracer_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_vest_urban.paa", //ODST_Chest
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_shoulders_urban.paa", //ODST_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_shoulders_urban.paa", //ODST_Right
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_M52_Marksman_Urban : 34thPRC_ArmourStandard_ODST2_M52_Marksman_Urban
-{
-    displayName="[34th] M52 Combat Armour 'Yandere' Marksman Urban";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_cqb_urban.paa", //CQB Left
-    	"34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_cqb_urban.paa", //CQB Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_marksman_urban.paa", //Marksman_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_marksman_urban.paa", //Marksman_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_vest_urban.paa", //ODST_Bracer_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_vest_urban.paa", //ODST_Bracer_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_vest_urban.paa", //ODST_Chest
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_shoulders_urban.paa", //ODST_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_shoulders_urban.paa", //ODST_Right
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_M52_CQB_Urban : 34thPRC_ArmourStandard_ODST2_M52_CQB_Urban
-{
-    displayName="[34th] M52 Combat Armour 'Yandere' CQB Urban";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_cqb_urban.paa", //CQB Left
-    	"34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_cqb_urban.paa", //CQB Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_marksman_urban.paa", //Marksman_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_marksman_urban.paa", //Marksman_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_vest_urban.paa", //ODST_Bracer_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_vest_urban.paa", //ODST_Bracer_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_vest_urban.paa", //ODST_Chest
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_shoulders_urban.paa", //ODST_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_shoulders_urban.paa", //ODST_Right
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_M52_Medic_Urban : 34thPRC_ArmourStandard_ODST2_M52_Urban
-{
-    displayName="[34th] M52 Combat Armour 'Yandere' (Corpsman) Urban";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-       "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_cqb_urban.paa", //CQB Left
-    	"34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_cqb_urban.paa", //CQB Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_marksman_urban.paa", //Marksman_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_marksman_urban.paa", //Marksman_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_vest_urban.paa", //ODST_Bracer_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_vest_urban.paa", //ODST_Bracer_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_vest_urban.paa", //ODST_Chest
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_shoulders_medic_urban.paa", //ODST_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_shoulders_medic_urban.paa", //ODST_Right
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_M52_B_Medic_Urban : 34thPRC_ArmourCustom_ODST2_Yandere_M52_B_Urban
-{
-	displayName="[34th] M52-B Combat Armour 'Yandere' (Corpsman Urban";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_cqb_urban.paa", //CQB Left
-    	"34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_cqb_urban.paa", //CQB Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_marksman_urban.paa", //Marksman_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_marksman_urban.paa", //Marksman_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_vest_medic_urban.paa", //ODST_Bracer_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_vest_medic_urban.paa", //ODST_Bracer_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_vest_medic_urban.paa", //ODST_Chest
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_shoulders_medic_urban.paa", //ODST_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_shoulders_medic_urban.paa", //ODST_Right
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_M52_C_Medic_Urban : 34thPRC_ArmourCustom_ODST2_Yandere_M52_C_Urban
-{
-	displayName="[34th] M52-C Combat Armour 'Yandere' Medic Urban";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_cqb_urban.paa", //CQB Left
-    	"34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_cqb_urban.paa", //CQB Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_marksman_urban.paa", //Marksman_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_marksman_urban.paa", //Marksman_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_vest_medic_urban.paa", //ODST_Bracer_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_vest_medic_urban.paa", //ODST_Bracer_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_vest_medic_urban.paa", //ODST_Chest
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_shoulders_medic_urban.paa", //ODST_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\urban\yandere_odst_shoulders_medic_urban.paa", //ODST_Right
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_M52_Arid : 34thPRC_ArmourCustom_ODST2_Yandere_M52_Urban
-{
-    displayName="[34th] M52 Combat Armour 'Yandere' Arid";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_cqb_arid.paa", //CQB Left
-    	"34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_cqb_arid.paa", //CQB Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_marksman_arid.paa", //Marksman_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_marksman_arid.paa", //Marksman_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_vest_arid.paa", //ODST_Bracer_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_vest_arid.paa", //ODST_Bracer_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_vest_arid.paa", //ODST_Chest
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_shoulders_arid.paa", //ODST_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_shoulders_arid.paa", //ODST_Right
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_M52_B_Arid : 34thPRC_ArmourCustom_ODST2_Yandere_M52_B_Urban
-{
-    displayName="[34th] M52-B Combat Armour 'Yandere' Arid";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_cqb_arid.paa", //CQB Left
-    	"34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_cqb_arid.paa", //CQB Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_marksman_arid.paa", //Marksman_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_marksman_arid.paa", //Marksman_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_vest_arid.paa", //ODST_Bracer_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_vest_arid.paa", //ODST_Bracer_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_vest_arid.paa", //ODST_Chest
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_shoulders_arid.paa", //ODST_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_shoulders_arid.paa", //ODST_Right
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_M52_C_Arid : 34thPRC_ArmourCustom_ODST2_Yandere_M52_C_Urban
-{
-    displayName="[34th] M52-C Combat Armour 'Yandere' Arid";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_cqb_arid.paa", //CQB Left
-    	"34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_cqb_arid.paa", //CQB Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_marksman_arid.paa", //Marksman_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_marksman_arid.paa", //Marksman_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_vest_arid.paa", //ODST_Bracer_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_vest_arid.paa", //ODST_Bracer_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_vest_arid.paa", //ODST_Chest
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_shoulders_arid.paa", //ODST_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_shoulders_arid.paa", //ODST_Right
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_M52_Marksman_Arid : 34thPRC_ArmourCustom_ODST2_Yandere_M52_Marksman_Urban
-{
-    displayName="[34th] M52 Combat Armour 'Yandere' Marksman Arid";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_cqb_arid.paa", //CQB Left
-    	"34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_cqb_arid.paa", //CQB Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_marksman_arid.paa", //Marksman_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_marksman_arid.paa", //Marksman_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_vest_arid.paa", //ODST_Bracer_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_vest_arid.paa", //ODST_Bracer_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_vest_arid.paa", //ODST_Chest
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_shoulders_arid.paa", //ODST_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_shoulders_arid.paa", //ODST_Right
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_M52_CQB_Arid : 34thPRC_ArmourCustom_ODST2_Yandere_M52_CQB_Urban
-{
-    displayName="[34th] M52 Combat Armour 'Yandere' CQB Arid";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_cqb_arid.paa", //CQB Left
-    	"34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_cqb_arid.paa", //CQB Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_marksman_arid.paa", //Marksman_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_marksman_arid.paa", //Marksman_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_vest_arid.paa", //ODST_Bracer_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_vest_arid.paa", //ODST_Bracer_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_vest_arid.paa", //ODST_Chest
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_shoulders_arid.paa", //ODST_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_shoulders_arid.paa", //ODST_Right
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_M52_Medic_Arid : 34thPRC_ArmourCustom_ODST2_Yandere_M52_Medic_Urban
-{
-    displayName="[34th] M52 Combat Armour 'Yandere' (Corpsman) Arid";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_cqb_arid.paa", //CQB Left
-    	"34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_cqb_arid.paa", //CQB Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_marksman_arid.paa", //Marksman_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_marksman_arid.paa", //Marksman_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_vest_medic_arid.paa", //ODST_Bracer_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_vest_medic_arid.paa", //ODST_Bracer_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_vest_medic_arid.paa", //ODST_Chest
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_shoulders_medic_arid.paa", //ODST_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_shoulders_medic_arid.paa", //ODST_Right
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_M52_B_Medic_Arid : 34thPRC_ArmourCustom_ODST2_Yandere_M52_B_Medic_Urban
-{
-	displayName="[34th] M52-B Combat Armour 'Yandere' (Corpsman) Arid";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_cqb_arid.paa", //CQB Left
-    	"34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_cqb_arid.paa", //CQB Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_marksman_arid.paa", //Marksman_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_marksman_arid.paa", //Marksman_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_vest_medic_arid.paa", //ODST_Bracer_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_vest_medic_arid.paa", //ODST_Bracer_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_vest_medic_arid.paa", //ODST_Chest
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_shoulders_medic_arid.paa", //ODST_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_shoulders_medic_arid.paa", //ODST_Right
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_M52_C_Medic_Arid : 34thPRC_ArmourCustom_ODST2_Yandere_M52_C_Medic_Urban
-{
-	displayName="[34th] M52-B Combat Armour 'Yandere' (Corpsman) Arid";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_cqb_arid.paa", //CQB Left
-    	"34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_cqb_arid.paa", //CQB Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_marksman_arid.paa", //Marksman_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_marksman_arid.paa", //Marksman_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_vest_medic_arid.paa", //ODST_Bracer_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_vest_medic_arid.paa", //ODST_Bracer_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_vest_medic_arid.paa", //ODST_Chest
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_shoulders_medic_arid.paa", //ODST_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\arid\yandere_odst_shoulders_medic_arid.paa", //ODST_Right
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_M52_Arctic : 34thPRC_ArmourCustom_ODST2_Yandere_M52_Urban
-{
-    displayName="[34th] M52 Combat Armour 'Yandere' Arctic";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_cqb_arctic.paa", //CQB Left
-    	"34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_cqb_arctic.paa", //CQB Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_marksman_arctic.paa", //Marksman_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_marksman_arctic.paa", //Marksman_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_vest_arctic.paa", //ODST_Bracer_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_vest_arctic.paa", //ODST_Bracer_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_vest_arctic.paa", //ODST_Chest
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_shoulders_arctic.paa", //ODST_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_shoulders_arctic.paa", //ODST_Right
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_M52_B_Arctic : 34thPRC_ArmourCustom_ODST2_Yandere_M52_B_Urban
-{
-    displayName="[34th] M52-B Combat Armour 'Yandere' Arctic";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_cqb_arctic.paa", //CQB Left
-    	"34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_cqb_arctic.paa", //CQB Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_marksman_arctic.paa", //Marksman_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_marksman_arctic.paa", //Marksman_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_vest_arctic.paa", //ODST_Bracer_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_vest_arctic.paa", //ODST_Bracer_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_vest_arctic.paa", //ODST_Chest
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_shoulders_arctic.paa", //ODST_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_shoulders_arctic.paa", //ODST_Right
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_M52_C_Arctic : 34thPRC_ArmourCustom_ODST2_Yandere_M52_C_Urban
-{
-    displayName="[34th] M52-C Combat Armour 'Yandere' Arctic";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_cqb_arctic.paa", //CQB Left
-    	"34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_cqb_arctic.paa", //CQB Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_marksman_arctic.paa", //Marksman_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_marksman_arctic.paa", //Marksman_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_vest_arctic.paa", //ODST_Bracer_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_vest_arctic.paa", //ODST_Bracer_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_vest_arctic.paa", //ODST_Chest
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_shoulders_arctic.paa", //ODST_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_shoulders_arctic.paa", //ODST_Right
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_M52_Marksman_Arctic : 34thPRC_ArmourCustom_ODST2_Yandere_M52_Marksman_Urban
-{
-    displayName="[34th] M52 Combat Armour 'Yandere' Marksman Arctic";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_cqb_arctic.paa", //CQB Left
-    	"34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_cqb_arctic.paa", //CQB Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_marksman_arctic.paa", //Marksman_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_marksman_arctic.paa", //Marksman_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_vest_arctic.paa", //ODST_Bracer_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_vest_arctic.paa", //ODST_Bracer_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_vest_arctic.paa", //ODST_Chest
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_shoulders_arctic.paa", //ODST_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_shoulders_arctic.paa", //ODST_Right
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_M52_CQB_Arctic : 34thPRC_ArmourCustom_ODST2_Yandere_M52_CQB_Urban
-{
-    displayName="[34th] M52 Combat Armour 'Yandere' CQB Arctic";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_cqb_arctic.paa", //CQB Left
-    	"34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_cqb_arctic.paa", //CQB Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_marksman_arctic.paa", //Marksman_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_marksman_arctic.paa", //Marksman_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_vest_arctic.paa", //ODST_Bracer_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_vest_arctic.paa", //ODST_Bracer_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_vest_arctic.paa", //ODST_Chest
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_shoulders_arctic.paa", //ODST_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_shoulders_arctic.paa", //ODST_Right
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_M52_Medic_Arctic : 34thPRC_ArmourCustom_ODST2_Yandere_M52_Medic_Urban
-{
-    displayName="[34th] M52 Combat Armour 'Yandere' (Corpsman) Arctic";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_cqb_arctic.paa", //CQB Left
-    	"34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_cqb_arctic.paa", //CQB Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_marksman_arctic.paa", //Marksman_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_marksman_arctic.paa", //Marksman_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_vest_medic_arctic.paa", //ODST_Bracer_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_vest_medic_arctic.paa", //ODST_Bracer_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_vest_medic_arctic.paa", //ODST_Chest
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_shoulders_medic_arctic.paa", //ODST_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_shoulders_medic_arctic.paa", //ODST_Right
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_M52_B_Medic_Arctic : 34thPRC_ArmourCustom_ODST2_Yandere_M52_B_Medic_Urban
-{
-	displayName="[34th] M52-B Combat Armour 'Yandere' (Corpsman) Arctic";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_cqb_arctic.paa", //CQB Left
-    	"34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_cqb_arctic.paa", //CQB Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_marksman_arctic.paa", //Marksman_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_marksman_arctic.paa", //Marksman_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_vest_medic_arctic.paa", //ODST_Bracer_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_vest_medic_arctic.paa", //ODST_Bracer_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_vest_medic_arctic.paa", //ODST_Chest
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_shoulders_medic_arctic.paa", //ODST_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_shoulders_medic_arctic.paa", //ODST_Right
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_M52_C_Medic_Arctic : 34thPRC_ArmourCustom_ODST2_Yandere_M52_C_Medic_Urban
-{
-	displayName="[34th] M52-C Combat Armour 'Yandere' (Corpsman) Arctic";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_cqb_arctic.paa", //CQB Left
-    	"34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_cqb_arctic.paa", //CQB Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_marksman_arctic.paa", //Marksman_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_marksman_arctic.paa", //Marksman_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_vest_medic_arctic.paa", //ODST_Bracer_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_vest_medic_arctic.paa", //ODST_Bracer_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_vest_medic_arctic.paa", //ODST_Chest
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_shoulders_medic_arctic.paa", //ODST_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\arctic\yandere_odst_shoulders_medic_arctic.paa", //ODST_Right
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_M52_Tropic : 34thPRC_ArmourCustom_ODST2_Yandere_M52_Urban
-{
-    displayName="[34th] M52 Combat Armour 'Yandere' Tropic";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_cqb_tropic.paa", //CQB Left
-    	"34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_cqb_tropic.paa", //CQB Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_marksman_tropic.paa", //Marksman_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_marksman_tropic.paa", //Marksman_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_vest_tropic.paa", //ODST_Bracer_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_vest_tropic.paa", //ODST_Bracer_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_vest_tropic.paa", //ODST_Chest
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_shoulders_tropic.paa", //ODST_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_shoulders_tropic.paa", //ODST_Right
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_M52_B_Tropic : 34thPRC_ArmourCustom_ODST2_Yandere_M52_B_Urban
-{
-    displayName="[34th] M52-B Combat Armour 'Yandere' Tropic";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_cqb_tropic.paa", //CQB Left
-    	"34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_cqb_tropic.paa", //CQB Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_marksman_tropic.paa", //Marksman_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_marksman_tropic.paa", //Marksman_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_vest_tropic.paa", //ODST_Bracer_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_vest_tropic.paa", //ODST_Bracer_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_vest_tropic.paa", //ODST_Chest
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_shoulders_tropic.paa", //ODST_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_shoulders_tropic.paa", //ODST_Right
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_M52_C_Tropic : 34thPRC_ArmourCustom_ODST2_Yandere_M52_C_Urban
-{
-    displayName="[34th] M52-C Combat Armour 'Yandere' Tropic";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_cqb_tropic.paa", //CQB Left
-    	"34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_cqb_tropic.paa", //CQB Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_marksman_tropic.paa", //Marksman_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_marksman_tropic.paa", //Marksman_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_vest_tropic.paa", //ODST_Bracer_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_vest_tropic.paa", //ODST_Bracer_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_vest_tropic.paa", //ODST_Chest
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_shoulders_tropic.paa", //ODST_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_shoulders_tropic.paa", //ODST_Right
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_M52_Marksman_Tropic : 34thPRC_ArmourCustom_ODST2_Yandere_M52_Marksman_Urban
-{
-    displayName="[34th] M52 Combat Armour 'Yandere' Marksman Tropic";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_cqb_tropic.paa", //CQB Left
-    	"34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_cqb_tropic.paa", //CQB Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_marksman_tropic.paa", //Marksman_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_marksman_tropic.paa", //Marksman_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_vest_tropic.paa", //ODST_Bracer_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_vest_tropic.paa", //ODST_Bracer_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_vest_tropic.paa", //ODST_Chest
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_shoulders_tropic.paa", //ODST_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_shoulders_tropic.paa", //ODST_Right
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_M52_CQB_Tropic : 34thPRC_ArmourCustom_ODST2_Yandere_M52_CQB_Urban
-{
-    displayName="[34th] M52 Combat Armour 'Yandere' CQB Tropic";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_cqb_tropic.paa", //CQB Left
-    	"34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_cqb_tropic.paa", //CQB Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_marksman_tropic.paa", //Marksman_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_marksman_tropic.paa", //Marksman_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_vest_tropic.paa", //ODST_Bracer_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_vest_tropic.paa", //ODST_Bracer_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_vest_tropic.paa", //ODST_Chest
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_shoulders_tropic.paa", //ODST_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_shoulders_tropic.paa", //ODST_Right
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_M52_Medic_Tropic : 34thPRC_ArmourCustom_ODST2_Yandere_M52_Medic_Urban
-{
-    displayName="[34th] M52 Combat Armour 'Yandere' (Corpsman) Tropic";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_cqb_tropic.paa", //CQB Left
-    	"34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_cqb_tropic.paa", //CQB Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_marksman_tropic.paa", //Marksman_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_marksman_tropic.paa", //Marksman_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_vest_medic_tropic.paa", //ODST_Bracer_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_vest_medic_tropic.paa", //ODST_Bracer_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_vest_medic_tropic.paa", //ODST_Chest
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_shoulders_medic_tropic.paa", //ODST_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_shoulders_medic_tropic.paa", //ODST_Right
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_M52_B_Medic_Tropic : 34thPRC_ArmourCustom_ODST2_Yandere_M52_B_Medic_Urban
-{
-    displayName="[34th] M52-B Combat Armour 'Yandere' (Corpsman) Tropic";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_cqb_tropic.paa", //CQB Left
-    	"34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_cqb_tropic.paa", //CQB Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_marksman_tropic.paa", //Marksman_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_marksman_tropic.paa", //Marksman_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_vest_medic_tropic.paa", //ODST_Bracer_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_vest_medic_tropic.paa", //ODST_Bracer_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_vest_medic_tropic.paa", //ODST_Chest
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_shoulders_medic_tropic.paa", //ODST_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_shoulders_medic_tropic.paa", //ODST_Right
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_M52_C_Medic_Tropic : 34thPRC_ArmourCustom_ODST2_Yandere_M52_C_Medic_Urban
-{
-    displayName="[34th] M52-B Combat Armour 'Yandere' (Corpsman) Tropic";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_cqb_tropic.paa", //CQB Left
-    	"34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_cqb_tropic.paa", //CQB Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_marksman_tropic.paa", //Marksman_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_marksman_tropic.paa", //Marksman_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_vest_medic_tropic.paa", //ODST_Bracer_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_vest_medic_tropic.paa", //ODST_Bracer_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_vest_medic_tropic.paa", //ODST_Chest
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_shoulders_medic_tropic.paa", //ODST_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_shoulders_medic_tropic.paa", //ODST_Right
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_M52_Woodland : 34thPRC_ArmourCustom_ODST2_Yandere_M52_Urban
-{
-    displayName="[34th] M52 Combat Armour 'Yandere' Woodland";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_cqb_woodland.paa", //CQB Left
-    	"34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_cqb_woodland.paa", //CQB Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_marksman_woodland.paa", //Marksman_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_marksman_woodland.paa", //Marksman_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_vest_woodland.paa", //ODST_Bracer_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_vest_woodland.paa", //ODST_Bracer_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_vest_woodland.paa", //ODST_Chest
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_shoulders_woodland.paa", //ODST_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_shoulders_woodland.paa", //ODST_Right
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_M52_B_Woodland : 34thPRC_ArmourCustom_ODST2_Yandere_M52_B_Urban
-{
-    displayName="[34th] M52-B Combat Armour 'Yandere' Woodland";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_cqb_woodland.paa", //CQB Left
-    	"34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_cqb_woodland.paa", //CQB Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_marksman_woodland.paa", //Marksman_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_marksman_woodland.paa", //Marksman_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_vest_woodland.paa", //ODST_Bracer_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_vest_woodland.paa", //ODST_Bracer_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_vest_woodland.paa", //ODST_Chest
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_shoulders_woodland.paa", //ODST_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_shoulders_woodland.paa", //ODST_Right
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_M52_C_Woodland : 34thPRC_ArmourCustom_ODST2_Yandere_M52_C_Urban
-{
-    displayName="[34th] M52-C Combat Armour 'Yandere' Woodland";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_cqb_woodland.paa", //CQB Left
-    	"34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_cqb_woodland.paa", //CQB Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_marksman_woodland.paa", //Marksman_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_marksman_woodland.paa", //Marksman_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_vest_woodland.paa", //ODST_Bracer_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_vest_woodland.paa", //ODST_Bracer_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_vest_woodland.paa", //ODST_Chest
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_shoulders_woodland.paa", //ODST_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_shoulders_woodland.paa", //ODST_Right
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_M52_Marksman_Woodland : 34thPRC_ArmourCustom_ODST2_Yandere_M52_Marksman_Urban
-{
-    displayName="[34th] M52 Combat Armour 'Yandere' Marksman Woodland";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_cqb_woodland.paa", //CQB Left
-    	"34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_cqb_woodland.paa", //CQB Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_marksman_woodland.paa", //Marksman_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_marksman_woodland.paa", //Marksman_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_vest_woodland.paa", //ODST_Bracer_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_vest_woodland.paa", //ODST_Bracer_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_vest_woodland.paa", //ODST_Chest
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_shoulders_woodland.paa", //ODST_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_shoulders_woodland.paa", //ODST_Right
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_M52_CQB_Woodland : 34thPRC_ArmourCustom_ODST2_Yandere_M52_CQB_Urban
-{
-    displayName="[34th] M52 Combat Armour 'Yandere' CQB Woodland";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_cqb_woodland.paa", //CQB Left
-    	"34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_cqb_woodland.paa", //CQB Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_marksman_woodland.paa", //Marksman_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_marksman_woodland.paa", //Marksman_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_vest_woodland.paa", //ODST_Bracer_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_vest_woodland.paa", //ODST_Bracer_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_vest_woodland.paa", //ODST_Chest
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_shoulders_woodland.paa", //ODST_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_shoulders_woodland.paa", //ODST_Right
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_M52_Medic_Woodland : 34thPRC_ArmourCustom_ODST2_Yandere_M52_Medic_Urban
-{
-    displayName="[34th] M52 Combat Armour 'Yandere' Medic Woodland";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_cqb_woodland.paa", //CQB Left
-    	"34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_cqb_woodland.paa", //CQB Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_marksman_woodland.paa", //Marksman_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_marksman_woodland.paa", //Marksman_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_vest_medic_woodland.paa", //ODST_Bracer_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_vest_medic_woodland.paa", //ODST_Bracer_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_vest_medic_woodland.paa", //ODST_Chest
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_shoulders_medic_woodland.paa", //ODST_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_shoulders_medic_woodland.paa", //ODST_Right
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_M52_B_Medic_Woodland : 34thPRC_ArmourCustom_ODST2_Yandere_M52_B_Medic_Urban
-{
-    displayName="[34th] M52-B Combat Armour 'Yandere' (Corpsman) Woodland";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_cqb_woodland.paa", //CQB Left
-    	"34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_cqb_woodland.paa", //CQB Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_marksman_woodland.paa", //Marksman_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_marksman_woodland.paa", //Marksman_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_vest_medic_woodland.paa", //ODST_Bracer_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_vest_medic_woodland.paa", //ODST_Bracer_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_vest_medic_woodland.paa", //ODST_Chest
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_shoulders_medic_woodland.paa", //ODST_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_shoulders_medic_woodland.paa", //ODST_Right
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
-    };
-};
-class 34thPRC_ArmourCustom_ODST2_Yandere_M52_C_Medic_Woodland : 34thPRC_ArmourCustom_ODST2_Yandere_M52_C_Medic_Urban
-{
-    displayName="[34th] M52-B Combat Armour 'Yandere' (Corpsman) Woodland";
-    author="Over Yandere";
-    hiddenSelectionsTextures[]=
-    {
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_cqb_woodland.paa", //CQB Left
-    	"34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_cqb_woodland.paa", //CQB Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_marksman_woodland.paa", //Marksman_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_marksman_woodland.paa", //Marksman_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_vest_medic_woodland.paa", //ODST_Bracer_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_vest_medic_woodland.paa", //ODST_Bracer_Right
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_vest_medic_woodland.paa", //ODST_Chest
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_shoulders_medic_woodland.paa", //ODST_Left
-        "34thPRC_ArmourCustom\data\odst2\yandere\woodland\yandere_odst_shoulders_medic_woodland.paa", //ODST_Right
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
-        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
-        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
-    };
-};
+#include "urban\config_weapons.hpp"
+#include "arid\config_weapons.hpp"
+#include "arctic\config_weapons.hpp"
+#include "tropic\config_weapons.hpp"
+#include "woodland\config_weapons.hpp"
