+++ src/Addons/34thPRC_ArmourCustom/data/odst2/yandere/tropic/config_weapons.hpp
@@ -1,0 +1,417 @@
+class 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_SilverVisor_Tropic : 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_SilverVisor_Urban
+{
+    displayName="[34th] CH252D Tropic 'Yandere' (Silver Visor)";
+    author="Over Yandere";
+    hiddenSelectionsTextures[]=
+    {
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_helmet_tropic.paa",
+        "34thPRC_ArmourStandard\data\odst2\visor\34th_ODST_Silvervisor.paa"
+    };
+};
+class 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_RedVisor_Tropic : 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_SilverVisor_Urban
+{
+	displayName="[34th] CH252D Tropic 'Yandere' (Red Visor)";
+    author="Over Yandere";
+    hiddenSelectionsTextures[]=
+    {
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_helmet_tropic.paa",
+        "34thPRC_ArmourStandard\data\odst2\visor\34th_ODST_Redvisor.paa"
+    };
+}
+class 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_GoldVisor_Tropic : 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_SilverVisor_Urban
+{
+	displayName="[34th] CH252D Tropic 'Yandere' (Gold Visor)";
+    author="Over Yandere";
+    hiddenSelectionsTextures[]=
+    {
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_helmet_tropic.paa",
+        "34thPRC_ArmourStandard\data\odst2\visor\34th_ODST_Goldvisor.paa"
+    };
+};
+class 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Tropic : 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_SilverVisor_Urban
+{
+	displayName="[34th] CH252D Tropic 'Yandere' (Custom Visor)";
+	author="Over Yandere";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_helmet_tropic.paa",
+		"34thPRC_ArmourCustom\data\odst2\yandere\visor\yandere_odst_visor.paa"
+	};
+};
+class 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_SilverVisor_Tropic_dp : 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_SilverVisor_Urban_dp
+{
+    displayName="[34th] CH252D Tropic 'Yandere' (Silver Visor)";
+    author="Over Yandere";
+	scope=0;
+    hiddenSelectionsTextures[]=
+    {
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_helmet_tropic.paa",// Helmet Texture
+        "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Depolarized_Visor_CA.paa" // Visor Texture
+    };
+     class ItemInfo: ItemInfo
+    {
+        uniformModel="MA_Armor\data\Helmets\Human_ODST\Reach_ODST_Helm_dp.p3d";
+        hiddenSelectionsTextures[]=
+        {
+            "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Helmet_CO.paa",
+            "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Depolarized_Visor_CA.paa"
+        };
+    };
+};
+class 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_RedVisor_Tropic_dp : 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_SilverVisor_Urban_dp
+{
+    displayName="[34th] CH252D Tropic 'Yandere' (Red Visor)";
+    author="Over Yandere";
+	scope=0;
+    hiddenSelectionsTextures[]=
+    {
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_helmet_tropic.paa",// Helmet Texture
+        "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Depolarized_Visor_CA.paa" // Visor Texture
+    };
+     class ItemInfo: ItemInfo
+    {
+        uniformModel="MA_Armor\data\Helmets\Human_ODST\Reach_ODST_Helm_dp.p3d";
+        hiddenSelectionsTextures[]=
+        {
+            "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Helmet_CO.paa",
+            "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Depolarized_Visor_CA.paa"
+        };
+    };
+};
+class 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_GoldVisor_Tropic_dp : 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_SilverVisor_Urban_dp
+{
+    displayName="[34th] CH252D Tropic 'Yandere' (Gold Visor)";
+    author="Over Yandere";
+	scope=0;
+    hiddenSelectionsTextures[]=
+    {
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_helmet_tropic.paa",// Helmet Texture
+        "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Depolarized_Visor_CA.paa" // Visor Texture
+    };
+     class ItemInfo: ItemInfo
+    {
+        uniformModel="MA_Armor\data\Helmets\Human_ODST\Reach_ODST_Helm_dp.p3d";
+        hiddenSelectionsTextures[]=
+        {
+            "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Helmet_CO.paa",
+            "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Depolarized_Visor_CA.paa"
+        };
+    };
+};
+class 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Tropic_dp : 34thPRC_ArmourCustom_ODST2_Yandere_CH252D_CustomVisor_Urban_dp
+{
+    displayName="[34th] CH252D Tropic 'Yandere' (Custom Visor)";
+    author="Over Yandere";
+	scope=0;
+    hiddenSelectionsTextures[]=
+    {
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_helmet_tropic.paa",// Helmet Texture
+        "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Depolarized_Visor_CA.paa" // Visor Texture
+    };
+     class ItemInfo: ItemInfo
+    {
+        uniformModel="MA_Armor\data\Helmets\Human_ODST\Reach_ODST_Helm_dp.p3d";
+        hiddenSelectionsTextures[]=
+        {
+            "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Helmet_CO.paa",
+            "MA_Armor\data\Helmets\Human_ODST\Color_Variants\MA_ODST_Depolarized_Visor_CA.paa"
+        };
+    };
+};
+
+class 34thPRC_ArmourCustom_ODST2_Yandere_M52_Tropic : 34thPRC_ArmourCustom_ODST2_Yandere_M52_Urban
+{
+    displayName="[34th] M52 Combat Armour 'Yandere' Tropic";
+    author="Over Yandere";
+    hiddenSelectionsTextures[]=
+    {
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_cqb_tropic.paa", //CQB Left
+    	"34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_cqb_tropic.paa", //CQB Right
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_marksman_tropic.paa", //Marksman_Left
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_marksman_tropic.paa", //Marksman_Right
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_vest_tropic.paa", //ODST_Bracer_Left
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_vest_tropic.paa", //ODST_Bracer_Right
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_vest_tropic.paa", //ODST_Chest
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_shoulders_tropic.paa", //ODST_Left
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_shoulders_tropic.paa", //ODST_Right
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
+        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
+        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
+        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
+        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
+    };
+};
+class 34thPRC_ArmourCustom_ODST2_Yandere_M52_B_Tropic : 34thPRC_ArmourCustom_ODST2_Yandere_M52_B_Urban
+{
+    displayName="[34th] M52-B Combat Armour 'Yandere' Tropic";
+    author="Over Yandere";
+    hiddenSelectionsTextures[]=
+    {
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_cqb_tropic.paa", //CQB Left
+    	"34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_cqb_tropic.paa", //CQB Right
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_marksman_tropic.paa", //Marksman_Left
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_marksman_tropic.paa", //Marksman_Right
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_vest_tropic.paa", //ODST_Bracer_Left
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_vest_tropic.paa", //ODST_Bracer_Right
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_vest_tropic.paa", //ODST_Chest
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_shoulders_tropic.paa", //ODST_Left
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_shoulders_tropic.paa", //ODST_Right
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
+        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
+        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
+        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
+        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
+    };
+};
+class 34thPRC_ArmourCustom_ODST2_Yandere_M52_C_Tropic : 34thPRC_ArmourCustom_ODST2_Yandere_M52_C_Urban
+{
+    displayName="[34th] M52-C Combat Armour 'Yandere' Tropic";
+    author="Over Yandere";
+    hiddenSelectionsTextures[]=
+    {
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_cqb_tropic.paa", //CQB Left
+    	"34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_cqb_tropic.paa", //CQB Right
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_marksman_tropic.paa", //Marksman_Left
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_marksman_tropic.paa", //Marksman_Right
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_vest_tropic.paa", //ODST_Bracer_Left
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_vest_tropic.paa", //ODST_Bracer_Right
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_vest_tropic.paa", //ODST_Chest
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_shoulders_tropic.paa", //ODST_Left
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_shoulders_tropic.paa", //ODST_Right
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
+        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
+        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
+        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
+        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
+    };
+};
+class 34thPRC_ArmourCustom_ODST2_Yandere_M52_Marksman_Tropic : 34thPRC_ArmourCustom_ODST2_Yandere_M52_Marksman_Urban
+{
+    displayName="[34th] M52 Combat Armour 'Yandere' Marksman Tropic";
+    author="Over Yandere";
+    hiddenSelectionsTextures[]=
+    {
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_cqb_tropic.paa", //CQB Left
+    	"34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_cqb_tropic.paa", //CQB Right
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_marksman_tropic.paa", //Marksman_Left
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_marksman_tropic.paa", //Marksman_Right
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_vest_tropic.paa", //ODST_Bracer_Left
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_vest_tropic.paa", //ODST_Bracer_Right
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_vest_tropic.paa", //ODST_Chest
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_shoulders_tropic.paa", //ODST_Left
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_shoulders_tropic.paa", //ODST_Right
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
+        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
+        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
+        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
+        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
+    };
+};
+class 34thPRC_ArmourCustom_ODST2_Yandere_M52_CQB_Tropic : 34thPRC_ArmourCustom_ODST2_Yandere_M52_CQB_Urban
+{
+    displayName="[34th] M52 Combat Armour 'Yandere' CQB Tropic";
+    author="Over Yandere";
+    hiddenSelectionsTextures[]=
+    {
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_cqb_tropic.paa", //CQB Left
+    	"34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_cqb_tropic.paa", //CQB Right
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_marksman_tropic.paa", //Marksman_Left
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_marksman_tropic.paa", //Marksman_Right
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_vest_tropic.paa", //ODST_Bracer_Left
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_vest_tropic.paa", //ODST_Bracer_Right
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_vest_tropic.paa", //ODST_Chest
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_shoulders_tropic.paa", //ODST_Left
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_shoulders_tropic.paa", //ODST_Right
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
+        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
+        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
+        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
+        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
+    };
+};
+class 34thPRC_ArmourCustom_ODST2_Yandere_M52_Medic_Tropic : 34thPRC_ArmourCustom_ODST2_Yandere_M52_Medic_Urban
+{
+    displayName="[34th] M52 Combat Armour 'Yandere' (Corpsman) Tropic";
+    author="Over Yandere";
+    hiddenSelectionsTextures[]=
+    {
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_cqb_tropic.paa", //CQB Left
+    	"34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_cqb_tropic.paa", //CQB Right
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_marksman_tropic.paa", //Marksman_Left
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_marksman_tropic.paa", //Marksman_Right
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_vest_medic_tropic.paa", //ODST_Bracer_Left
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_vest_medic_tropic.paa", //ODST_Bracer_Right
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_vest_medic_tropic.paa", //ODST_Chest
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_shoulders_medic_tropic.paa", //ODST_Left
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_shoulders_medic_tropic.paa", //ODST_Right
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
+        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
+        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
+        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
+        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
+    };
+};
+class 34thPRC_ArmourCustom_ODST2_Yandere_M52_B_Medic_Tropic : 34thPRC_ArmourCustom_ODST2_Yandere_M52_B_Medic_Urban
+{
+    displayName="[34th] M52-B Combat Armour 'Yandere' (Corpsman) Tropic";
+    author="Over Yandere";
+    hiddenSelectionsTextures[]=
+    {
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_cqb_tropic.paa", //CQB Left
+    	"34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_cqb_tropic.paa", //CQB Right
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_marksman_tropic.paa", //Marksman_Left
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_marksman_tropic.paa", //Marksman_Right
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_vest_medic_tropic.paa", //ODST_Bracer_Left
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_vest_medic_tropic.paa", //ODST_Bracer_Right
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_vest_medic_tropic.paa", //ODST_Chest
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_shoulders_medic_tropic.paa", //ODST_Left
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_shoulders_medic_tropic.paa", //ODST_Right
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
+        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
+        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
+        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
+        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
+    };
+};
+class 34thPRC_ArmourCustom_ODST2_Yandere_M52_C_Medic_Tropic : 34thPRC_ArmourCustom_ODST2_Yandere_M52_C_Medic_Urban
+{
+    displayName="[34th] M52-B Combat Armour 'Yandere' (Corpsman) Tropic";
+    author="Over Yandere";
+    hiddenSelectionsTextures[]=
+    {
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_cqb_tropic.paa", //CQB Left
+    	"34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_cqb_tropic.paa", //CQB Right
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_marksman_tropic.paa", //Marksman_Left
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_marksman_tropic.paa", //Marksman_Right
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_vest_medic_tropic.paa", //ODST_Bracer_Left
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_vest_medic_tropic.paa", //ODST_Bracer_Right
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_vest_medic_tropic.paa", //ODST_Chest
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_shoulders_medic_tropic.paa", //ODST_Left
+        "34thPRC_ArmourCustom\data\odst2\yandere\tropic\yandere_odst_shoulders_medic_tropic.paa", //ODST_Right
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMLeft
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //ChestPMRight
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //ChestPouch
+        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //LShoulderRadio
+        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //RShoulderRadio
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //StomachPouch
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMLeft
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //TorsoPMRight
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //TorsoPouch
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistBack
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGLeft
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistGRight
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPLeft
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_A_CO.paa", //WaistPRight
+        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRLeft
+        "MA_Armor\data\Backpacks\ODST_Rucksack\Attachments\Backpack_Radio_ODST_co.paa", //WaistRRight
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSLeft
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //WaistSRight
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa", //LegPouchL
+        "MA_Armor\data\vests\Marine_Pouches\data\MA_TrooperExtras_B_CO.paa"  //LegPouchR
+    };
+};
